tích hợp level, kinh nghiệm, điểm kỹ năng lên cửa sổ trạng thái + tích hợp điểm kinh nghiệm khi giết quái

+++ Assets/Scripts/Player/ExpBar.cs
@@ -9,7 +9,7 @@ public class ExpBar : MonoBehaviour
     public TextMeshProUGUI levelText;
     public TextMeshProUGUI expAmountText;
 
-    public void SetExp(float currentExp, float expToNextLevel)
+    public void SetExp(int currentExp, int expToNextLevel)
     {
         if (slider != null)
         {
